fix(ICE): Fix candidate priority
It was always 0. Usage not implemented at this point, but maybe in the future.

+++ src/media/delivery/ice.h
@@ -60,7 +60,7 @@ private:
                                                            const QList<std::shared_ptr<ICEInfo> > &remote,
                                                            bool controller);
 
-  int pairPriority(int controllerCandidatePriority, int controlleeCandidatePriority) const;
+  uint64_t pairPriority(int controllerCandidatePriority, int controlleeCandidatePriority) const;
 
   bool sameCandidates(std::vector<std::shared_ptr<ICEPair>> newCandidates,
                       std::vector<std::shared_ptr<ICEPair>> oldCandidates);
